pyproject.toml
[project]
name = "gas"
version = "0.1.0"
description = "Git AI Sidekick - A CLI tool for AI-powered Git operations"
authors = []
dependencies = [
    "click>=8.1.0",
    "huggingface-hub>=0.20.0",
    "python-dotenv>=1.0.0",
    "gitpython>=3.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.1",
]
requires-python = ">=3.13"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
gas = "gas.cli:cli"

[tool.ruff]
line-length = 100
target-version = "py39"
